requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alembic==1.10.4
    # via -r requirements.in
anyio==3.6.2
    # via starlette
astor==0.8.1
    # via wemake-python-styleguide
astroid==2.15.4
    # via pylint
async-timeout==4.0.2
    # via redis
asyncpg==0.27.0
    # via -r requirements.in
attrs==23.1.0
    # via
    #   flake8-bugbear
    #   flake8-eradicate
    #   wemake-python-styleguide
bandit==1.7.2
    # via
    #   -r requirements.in
    #   flake8-bandit
build==0.10.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via
    #   pip-tools
    #   uvicorn
darglint==1.8.1
    # via wemake-python-styleguide
dependency-injector==4.41.0
    # via -r requirements.in
dill==0.3.6
    # via pylint
distlib==0.3.6
    # via virtualenv
docutils==0.19
    # via restructuredtext-lint
eradicate==2.2.0
    # via flake8-eradicate
exceptiongroup==1.1.1
    # via pytest
fastapi==0.95.1
    # via -r requirements.in
filelock==3.12.0
    # via virtualenv
flake8==3.9.1
    # via
    #   -r requirements.in
    #   flake8-bandit
    #   flake8-broken-line
    #   flake8-bugbear
    #   flake8-commas
    #   flake8-comprehensions
    #   flake8-debugger
    #   flake8-docstrings
    #   flake8-eradicate
    #   flake8-for-pycharm
    #   flake8-isort
    #   flake8-polyfill
    #   flake8-quotes
    #   flake8-rst-docstrings
    #   flake8-string-format
    #   pep8-naming
    #   wemake-python-styleguide
flake8-bandit==2.1.2
    # via
    #   -r requirements.in
    #   wemake-python-styleguide
flake8-broken-line==0.4.0
    # via wemake-python-styleguide
flake8-bugbear==22.12.6
    # via wemake-python-styleguide
flake8-commas==2.1.0
    # via wemake-python-styleguide
flake8-comprehensions==3.12.0
    # via wemake-python-styleguide
flake8-debugger==4.1.2
    # via wemake-python-styleguide
flake8-docstrings==1.7.0
    # via wemake-python-styleguide
flake8-eradicate==1.4.0
    # via wemake-python-styleguide
flake8-for-pycharm==0.4.1
    # via -r requirements.in
flake8-isort==4.2.0
    # via
    #   -r requirements.in
    #   wemake-python-styleguide
flake8-polyfill==1.0.2
    # via
    #   flake8-bandit
    #   pep8-naming
flake8-quotes==3.3.2
    # via wemake-python-styleguide
flake8-rst-docstrings==0.2.7
    # via wemake-python-styleguide
flake8-string-format==0.3.0
    # via wemake-python-styleguide
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via bandit
greenlet==2.0.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.5.0
    # via -r requirements.in
identify==2.5.24
    # via pre-commit
idna==3.4
    # via anyio
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via
    #   flake8-isort
    #   pylint
lazy-object-proxy==1.9.0
    # via astroid
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via mako
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
nodeenv==1.7.0
    # via pre-commit
orjson==3.8.12
    # via -r requirements.in
packaging==23.1
    # via
    #   build
    #   pytest
pbr==5.11.1
    # via stevedore
pep8-naming==0.12.1
    # via wemake-python-styleguide
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.5.0
    # via
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.15.0
    # via -r requirements.in
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-bandit
    #   flake8-debugger
pydantic[dotenv]==1.10.7
    # via
    #   -r requirements.in
    #   fastapi
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==2.3.1
    # via flake8
pygments==2.15.1
    # via
    #   flake8-rst-docstrings
    #   wemake-python-styleguide
pylint==2.17.4
    # via
    #   -r requirements.in
    #   flake8-for-pycharm
pyproject-hooks==1.0.0
    # via build
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.21.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via pydantic
pyyaml==6.0
    # via
    #   bandit
    #   pre-commit
redis==4.5.4
    # via -r requirements.in
restructuredtext-lint==1.4.0
    # via flake8-rst-docstrings
six==1.16.0
    # via dependency-injector
smmap==5.0.0
    # via gitdb
sniffio==1.3.0
    # via anyio
snowballstemmer==2.2.0
    # via pydocstyle
sqlalchemy[asyncio]==2.0.12
    # via
    #   -r requirements.in
    #   alembic
starlette==0.26.1
    # via fastapi
stevedore==5.0.0
    # via bandit
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   build
    #   pylint
    #   pyproject-hooks
    #   pytest
tomlkit==0.11.8
    # via pylint
typing-extensions==4.5.0
    # via
    #   alembic
    #   astroid
    #   pydantic
    #   sqlalchemy
    #   wemake-python-styleguide
uvicorn==0.22.0
    # via -r requirements.in
#uvloop==0.17.0
    # via -r requirements.in
virtualenv==20.23.0
    # via pre-commit
wemake-python-styleguide==0.16.1
    # via -r requirements.in
wheel==0.40.0
    # via pip-tools
wrapt==1.15.0
    # via astroid
kafka-python==2.0.2

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
